refactor: move welcome-routing helpers into the beta.10 version file
Per review: migrated_from_pre_0_4_0 and should_welcome_to_release are
beta.10-specific, not general migration facilities, so they belong with the
version that uses them rather than in version/mod.rs. Behavior and the
welcome_routing test are unchanged (the test moves with them).
